controllers.flights: use SkyLinesPolyEncoder instead of cGPolyEncoder
The SkyLinesPolyEncoder is a fork of cGPolyEncoder. It adds the posibility
to simplicate a trace by a maximum point-to-point distance. This is needed
to keep uniformly distributed points in time for the barogram graph.

Additionally the trace classification and google style encoding is splitted
up so the map display and barogram get exactly the same points to display.