controllers/error: Fixed DetachedInstanceError throws
When a HTTPError was raise the current DBSession was closed and the
request.identity['user'] object detached from the session. Since the topbar
template uses the club property of the object with deferred loading the
object needs to be attached to a new DBSession before it can be used again.