model.User: add attribute eye_candy, disables OpenLayers fading
According to a benchmark done with Chromium 20 and "perf stat", fading
map tiles in adds 50% CPU cycles.  This is a huge waste, and delays
the page by a unacceptable amount of time.  To allow those who want it
still see this eye candy, this commit adds a per-user option to
re-enable the fading.