users: Allow users to change their email address properly
Previously the uniqueness of the email address wasn't protected in python.
Using the UniqueValue validator isn't possible because if the email address
isn't changed the supplied email address would already exist in the
database and cause an error. The new UniqueValueUnless validator is using
an additional filter that filters out the current user from the uniqueness
check.