public/js/flight: fix hoverTolerance when searching for flight track
Setting distance to a large initial value failed if the vector was
very long and therefore the hovering tolerance was too large (at times
larger than the whole map window).

We now calculate the real map resolution at the mouse cursor and use
this value for calculating the maximum allowable hover distance.