controllers: Fixed sprox fields
If Field() is used the first parameter should be a class instead of an
instance, otherwise the user has the entire responsibility of configuring
the widget properly. That was done wrong in this case and should be
discouraged.

Instead of creating an instance of the widget class the parameters are set
through the __field_widget_args__ attribute of the ViewBase class.

This fixes the validation for the affected forms.