model: add ContestLeg table
This table is intended to store the legs of a optimized contest result.
It overcomes some shortcomings of the traces table and will therefore
replace it.

The main benefits are additional columns to store leg statistics like
circling duration or total height loss. These values cannot be calculated
without the complete flight path and are therefore to be evaluated on
flight analysis.